aider: Added a flag `--dbfile` to specify the DuckDB database file with a default value of `plummet.db` in the `main.go` file.
# Aider chat conversation:

USER: We need to use `DuckDB` as our database. We want to add a flag to the file we should use for the duckdb database file. The default should be `plummet.db` in the current directory.

ASSISTANT: To add a flag for the DuckDB database file with a default value of `plummet.db`, we need to modify the `main.go` file. Here's the diff for the changes:

```diff
--- cmd/plummet/main.go
+++ cmd/plummet/main.go
@@ ... @@
 func main() {
 	app := &cli.App{
 		Name:  "plummet",
 		Usage: "A build system that runs SQL against a database",
+		Flags: []cli.Flag{
+			&cli.StringFlag{
+				Name:    "dbfile",
+				Value:   "plummet.db",
+				Usage:   "DuckDB database file",
+				Aliases: []string{"d"},
+			},
+		},
 		Action: func(c *cli.Context) error {
 			file, err := os.ReadFile("plummet.yml")
 			if err != nil {
 				log.Fatalf("Unable to read plummet.yml: %v", err)
 			}

 			var plummetFile PlummetFile
 			err = yaml.Unmarshal(file, &plummetFile)
 			if err != nil {
 				log.Fatalf("Unable to parse plummet.yml: %v", err)
 			}

+			dbFile := c.String("dbfile")
+
 			if c.Args().Len() > 0 {
 				targetName := c.Args().First()
 				visited := make(map[string]bool)
 				err := executeTarget(targetName, &plummetFile, visited)
 				if err != nil {
 					log.Fatal(err)
 				}
 			} else {
 				fmt.Println("Available targets:")
 				for target := range plummetFile.Targets {
 					fmt.Println(target)
 				}
 			}

 			return nil
 		},
 	}
 	err := app.Run(os.Args)
 	if err != nil {
 		log.Fatal(err)
 	}
 }
```

This diff adds a new flag `--dbfile` (with the short alias `-d`) to the CLI app, which specifies the DuckDB database file to use, with a default value of `plummet.db`. The variable `dbFile` is then set to the value of this flag, which can be used later in the code to connect to the DuckDB database.