aider: Add `Config` field to `Target` struct.
# Aider chat conversation:

USER: Lets update the `Target`

ASSISTANT: Could you please specify the changes you would like to make to the `Target` struct?

USER: Update the `Target` to support a `config` element that is a map[string]interface{}

ASSISTANT: Here is the diff to update the `Target` struct in `cmd/plummet/main.go` to include a `Config` field of type `map[string]interface{}`:

```diff
--- cmd/plummet/main.go
+++ cmd/plummet/main.go
@@ ... @@
 type Target struct {
 	Output string   `yaml:"output"`
 	SQL    string   `yaml:"sql"`
 	Deps   []string `yaml:"deps"`
+	Config map[string]interface{} `yaml:"config"`
 }
```

This change adds a new field `Config` to the `Target` struct, which allows for a map of configurations to be associated with each target.