Quick update about the output file
We don't use the output file as a dependency, but I think we could
potentially. I think the more important thing may be avoiding running
some part of the job. This could be a built-in concept, but a better
tactic might simply be to allow dyanmic parameters to be passed in.

For example, if your job is meant to pull in data every day, you need
to pass in some concept of "now" or "today" to the SQL. Some functions
that make dealing with time or storing output from other queries would
likely work.

The time aspect is reasonably understood, but if someone is using
autoincrementing ID fields and they want to get everything larger than
some ID, we could allow storing a "last_id". The SQL job could have a
simple scratch table as well with a set schema that allows someone to
easily do:

```
INSERT INTO target_metadata (key, value)
VALUES ('last_id', (SELECT max(id) FROM some_table));
```

You could then join with the target_metedata table in your SQL or
call a function that would query under the hood to add some
convenience.